React query templates
