+++ package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.4.2",
+    "@lukemorales/query-key-factory": "^1.3.4",
     "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-slot": "^1.1.0",
     "@sentry/react": "^8.4.0",
