+++ src/config/api/index.ts
@@ -1,3 +1,4 @@
+export * from "./api.types";
 export * from "./axios";
 export * from "./handleAxiosFieldErrors";
 export * from "./interceptors";
