+++ src/router/index.ts
@@ -1,3 +1,5 @@
-export * from "./Router";
-export * from "./constants";
+export * from "./router";
 export * from "./components";
+export * from "./constants";
+export * from "./layouts";
+export * from "./types";
