+++ templates/domains/api/index.ts.hbs
@@ -1,1 +1,1 @@
-export * from "./{{camelCase name}}";
+export * from "./{{kebabCase name}}";
