+++ templates/domains/context/context.hbs
@@ -1,0 +1,5 @@
+import * as React from "react"
+
+export const use{{pascalCase name}}Context: React.FC = () => {
+  return <div>This is an Component Example.</div>
+}
