+++ templates/domains/context/index.ts.hbs
@@ -1,1 +1,1 @@
-export * from "./useExampleContext";
+export * from "./use{{pascalCase name}}Context";
