+++ templates/domains/queries/index.ts.hbs
@@ -1,1 +1,1 @@
-export * from "./useSomethingQueryExample";
+export * from "./{{kebabCase name}}";
