+++ templates/domains/queries/query-example.tsx.hbs
@@ -1,5 +1,133 @@
-import * as React from "react"
+import { createQueryKeys } from "@lukemorales/query-key-factory";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
-export const useSomethingQueryExample = () => {
-  return <div>This is a useQueryExample component.</div>
-}
+import type {
+  Base{{pascalCase name}},
+  Base{{pascalCase name}}ListParams,
+} from "../api/{{kebabCase name}}";
+import {
+  create{{pascalCase name}},
+  delete{{pascalCase name}},
+  get{{pascalCase name}}Detail,
+  get{{pascalCase name}}List,
+  update{{pascalCase name}},
+} from "../api/{{kebabCase name}}";
+
+export const {{camelCase name}}Keys = createQueryKeys("{{camelCase name}}", {
+  list: null,
+  detail: ({{camelCase name}}Id: string) => ({
+    queryKey: [{{camelCase name}}Id],
+    queryFn: () => get{{pascalCase name}}Detail({{camelCase name}}Id),
+  }),
+});
+
+export const use{{pascalCase name}} = () => {
+  const queryClient = useQueryClient();
+
+  const use{{pascalCase name}}List = (params: Base{{pascalCase name}}ListParams) =>
+    useQuery({
+      ...{{camelCase name}}Keys.list,
+      queryFn: () => get{{pascalCase name}}List(params),
+    });
+
+  const use{{pascalCase name}}Detail = ({{camelCase name}}Id: string) =>
+    useQuery({{camelCase name}}Keys.detail({{camelCase name}}Id));
+
+  const usePrefetch{{pascalCase name}}Detail = ({{camelCase name}}Id: string) =>
+    queryClient.prefetchQuery({{camelCase name}}Keys.detail({{camelCase name}}Id));
+
+  const useCreate{{pascalCase name}} = useMutation({
+    mutationFn: create{{pascalCase name}},
+    onSuccess: (new{{pascalCase name}}) => {
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail(new{{pascalCase name}}.id).queryKey,
+        new{{pascalCase name}},
+      );
+
+      void queryClient.invalidateQueries({
+        queryKey: {{camelCase name}}Keys._def,
+      });
+    },
+  });
+
+  const useUpdate{{pascalCase name}} = useMutation({
+    mutationFn: update{{pascalCase name}},
+    onMutate: async (new{{pascalCase name}}) => {
+      // Cancel any outgoing refetches (so they don't overwrite our optimistic update)
+      await queryClient.cancelQueries({{camelCase name}}Keys.detail(new{{pascalCase name}}.id));
+
+      // Snapshot the previous value
+      const previousTasks = queryClient.getQueryData(
+        {{camelCase name}}Keys.detail(new{{pascalCase name}}.id).queryKey,
+      );
+
+      // Optimistically update to the new value
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail(new{{pascalCase name}}.id).queryKey,
+        new{{pascalCase name}},
+      );
+
+      // Return a context object with the snapshotted value
+      return { previousTasks };
+    },
+    onError: (_err, {{camelCase name}}, context) => {
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail({{camelCase name}}.id).queryKey,
+        context?.previousTasks,
+      );
+    },
+    onSuccess: (new{{pascalCase name}}) => {
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail(new{{pascalCase name}}.id).queryKey,
+        new{{pascalCase name}},
+      );
+
+      void queryClient.invalidateQueries({
+        queryKey: {{camelCase name}}Keys._def,
+      });
+    },
+  });
+
+  const useDelete{{pascalCase name}} = useMutation({
+    mutationFn: delete{{pascalCase name}},
+    onMutate: async ({{camelCase name}}Id) => {
+      // Cancel any outgoing refetches (so they don't overwrite our optimistic update)
+      await queryClient.cancelQueries({{camelCase name}}Keys.detail({{camelCase name}}Id));
+
+      // Snapshot the previous value
+      const previousTasks = queryClient.getQueryData(
+        {{camelCase name}}Keys.detail({{camelCase name}}Id).queryKey,
+      );
+
+      // Optimistically update to the new value
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail({{camelCase name}}Id).queryKey,
+        (old: Base{{pascalCase name}}[]) =>
+          old.filter((t: Base{{pascalCase name}}) => t.id !== {{camelCase name}}Id),
+      );
+
+      // Return a context object with the snapshotted value
+      return { previousTasks };
+    },
+    onError: (_err, {{camelCase name}}Id, context) => {
+      queryClient.setQueryData(
+        {{camelCase name}}Keys.detail({{camelCase name}}Id).queryKey,
+        context?.previousTasks,
+      );
+    },
+    onSuccess: () => {
+      void queryClient.invalidateQueries({
+        queryKey: {{camelCase name}}Keys._def,
+      });
+    },
+  });
+
+  return {
+    use{{pascalCase name}}List,
+    use{{pascalCase name}}Detail,
+    usePrefetch{{pascalCase name}}Detail,
+    useCreate{{pascalCase name}},
+    useUpdate{{pascalCase name}},
+    useDelete{{pascalCase name}},
+  };
+};
