+++ templates/domains/sections/index.ts.hbs
@@ -1,1 +1,1 @@
-export * from "./ExampleSection";
+export * from "./SectionExample";
